requirements.txt
# FastAPI и веб-сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Blockchain (mock, можно заменить на реальный web3)
web3==6.15.1

# Utilities
python-dotenv==1.0.0

